fix function argument type in __acl_get_file.h

+++ libacl/__acl_get_file.h
@@ -1,4 +1,4 @@
 acl_t __acl_get_file(const char *, acl_type_t ,
 	             ssize_t (*)(const char *, const char *,
 				   void *, size_t),
-	             ssize_t (*)(const char*, struct stat*));
+	             int (*)(const char*, struct stat*));
